Set default command for `fzf` via `fd`/`rg`
A combination of 3 of my favorite new tools ;)

If `fzf` is installed on `$PATH`, then attempt to set a better/more
performant default command by setting `$FZF_DEFAULT_COMMAND`. Prefer
using `fd` if present and fall back to `rg` if present. Both are amazing
and fast, although `fd` is designed for filesystem finds, so prefer
that.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>