Fix `cf` function to support optional directory arg
The syntax of `fd` is slightly different from `find` in that the
directory argument is required and *first* with `find` and optional and
*second* with `fd`.

This fix adds a catch-all pattern `.` for `fd` so that the second
argument is now interpreted as the directory. Note that by default `.`
is provided which means it's search the current working directory.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>